requirements.txt
appdirs==1.4.3
dj-database-url==0.4.2
Django==1.10.6
django-cors-headers==2.0.2
djangorestframework==3.6.2
djangorestframework-jwt==1.10.0
httpie==0.9.9
ordered-set==2.0.2
packaging==16.8
psycopg2==2.7
Pygments==2.2.0
PyJWT==1.5.0
PyLaTeX==1.2.0
pyparsing==2.2.0
requests==2.13.0
six==1.10.0
